% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_dfa.R
\name{cluster_dfa}
\alias{cluster_dfa}
\title{Clustering function to process DFA results.}
\usage{
cluster_dfa(data_dfa, species_sub, nboot = 500, min_stability = 0.5)
}
\arguments{
\item{data_dfa}{A \code{fit_dfa} object. Results of \code{fit_dfa()}.}

\item{species_sub}{A \code{data.frame} of species names. It should be provided with species in row, the first column for complete species names and the second column for species names' codes.}

\item{nboot}{An \code{integer}. Number of bootstrap iterations for clustering. Default is 500.}

\item{min_stability}{A \code{numeric} value between 0 and 1. Minimum cluster stability. If the stability of one cluster is below this value, the cluster is dilluted in the remaining clusters. Default is 0.5. Be aware that if the value is set below 0.5, this enables to keep clusters with low stability that may not be relevant.}
}
\value{
A \code{list} of three objects: \code{group_dfa} clustering output, \code{tmbObj} updated output of \code{MakeADFun()}, \code{sdRep} updated output of TMB optimisation.
}
\description{
Clustering function to process DFA results.
}
\examples{
\dontrun{
data(species_ts_mat)
data(species_uncert_ts_mat)

data_ready_dfa <- prepare_data(data_ts = species_ts_mat,data_ts_se = species_uncert_ts_mat,
se_log = TRUE, perc_replace = 0.01)

dfa_result <- fit_dfa(data_ts = data_ready_dfa$data_ts,data_ts_se = data_ready_dfa$data_ts_se,
min_year = data_ready_dfa$min_year, max_year = data_ready_dfa$max_year, species_name_ordre = data_ready_dfa$species_name_ordre,
species_sub = species_name, nfac = 0, mintrend = 1, maxtrend = 5, AIC = TRUE,
center_option = 1, silent = TRUE, control = list())

data(species_name)

cluster_result <- cluster_dfa(data_dfa = dfa_result, species_sub = species_name, nboot = 500)
}
}
