% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_dfa.R
\name{core_dfa}
\alias{core_dfa}
\title{Core function running the DFA and compute the AIC for a given number of latent trends (internal function)}
\usage{
core_dfa(
  data_ts,
  data_ts_se,
  nfac,
  AIC = TRUE,
  silent = TRUE,
  control = list(),
  center_option = 1
)
}
\arguments{
\item{data_ts}{A \code{matrix} of species time series. It should be provided as a \code{matrix} with species time-series in row, species' codes as row names and years as column names.}

\item{data_ts_se}{A \code{matrix} of species of uncertainty (e.g. standard error) of species time series. It should be provided as a \code{matrix} with log species uncertainty time-series in row, species' codes as row names and years as column names.}

\item{nfac}{An \code{integer}. Number of trends for the DFA.}

\item{AIC}{A \code{logical} value. \code{TRUE} computes and displays the AIC, \code{FALSE} does not. Default is TRUE.}

\item{silent}{A \code{logical} value. \code{TRUE} silences \code{MakeADFun()}, \code{FALSE} does not. Default is TRUE.}

\item{control}{A \code{list}. Control option for \code{MakeADFun()}. Default is list().}

\item{center_option}{An \code{integer}. Option to handle time-series centered according to the first year (\code{center_option} = 0) or mean centred (default, \code{center_option} = 1).}
}
\value{
A \code{list} of 13 objects: \code{tmbObj} the output of \code{MakeADFun()}, \code{tmbOpt} the optimisation from \code{tmbObj}, \code{data_ts} the processed \code{matrix} of species time series, \code{data_ts_se} the processed \code{matrix} of time series uncertainty, \code{data_ts_save} a \code{data.frame} of input species time series, \code{data_ts_save_long} a \code{data.frame} of input species time series in long format, \code{data_ts_se_save} a \code{data.frame} of input species uncertainty time series, \code{ny} the number of time series, \code{nT} the number of time steps, \code{aic} the value of the AIC, \code{conv} the result of the convergence check, \code{sdRep_test} the summary of the TMB optimisation output, \code{sdRep_test_all} the complete TMB optimisation output.
}
\description{
Core function running the DFA and compute the AIC for a given number of latent trends (internal function)
}
\examples{
\dontrun{
core_dfa(data_ts=data_ts, data_ts_se=data_ts_se, nfac=i, silent = silent, control = control, center_option = 1)
}
}
