% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_check_prepare.R
\name{data_check_prepare}
\alias{data_check_prepare}
\title{Function to check data before DFA}
\usage{
data_check_prepare(
  data_ts,
  data_ts_se,
  se_log = TRUE,
  is_mean_centred = TRUE,
  min_year_sc = NULL
)
}
\arguments{
\item{data_ts}{A \code{matrix}. Dataset of species time series. It should be provided as a matrix with species time-series in row, species' codes as row names and years as column names.}

\item{data_ts_se}{A \code{matrix}. Dataset of uncertainty (e.g. standard error) of species time series. It should be provided as a matrix with log uncertainty of species time-series in row, species' codes as row names and years as column names.}

\item{se_log}{A \code{logical} value. \code{TRUE} if \code{data_ts_se} is provided with log values. \code{FALSE} if it is provided with not log values. Default is \code{TRUE}.}

\item{is_mean_centred}{A \code{logical} value. \code{TRUE} if \code{data_ts} is provided with mean-centred values. \code{FALSE} if it is provided with non mean-centred values. Default is \code{TRUE}.}

\item{min_year_sc}{An \code{integer}. First year of the period to select for rescaling if \code{is_mean_centred} = \code{FALSE}.}
}
\value{
A \code{list} of five objects: \code{min_year} an \code{integer} corresponding to the first year of time-series, \code{max_year} an \code{integer} corresponding to the last year of time-series, \code{species_name_ordre} a \verb{character vector} of species code following the ordre of \code{data_ts}, \code{data_ts} a \code{matrix} with ready-to-use time-series for DFA (\code{fun_make_dfa()}), \code{data_ts_se} a \code{matrix} with ready-to-use uncertainty for DFA (\code{fun_make_dfa()}).
}
\description{
Function to check data before DFA
}
\examples{
data(species_ts_mat)
data(species_uncert_ts_mat)
species_ts_mat[species_ts_mat==0] <- NA
data_ready_dfa <- data_check_prepare(data_ts = species_ts_mat,data_ts_se = species_uncert_ts_mat,
se_log = FALSE,is_mean_centred = FALSE, min_year_sc=2000)
}
