% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dfa.R
\name{fit_dfa}
\alias{fit_dfa}
\title{Run DFA}
\usage{
fit_dfa(
  data_ts,
  data_ts_se,
  min_year,
  max_year,
  species_name_ordre,
  species_sub,
  nfac = 0,
  mintrend = 1,
  maxtrend = 5,
  AIC = TRUE,
  center_option = 1,
  silent = TRUE,
  control = list()
)
}
\arguments{
\item{data_ts}{A \code{matrix} of species time series. It should be provided as a \code{matrix} with species time-series in row, species' codes as row names and years as column names.}

\item{data_ts_se}{A \code{matrix} of species of uncertainty (e.g. standard error) of species time series. It should be provided as a \code{matrix} with log species uncertainty time-series in row, species' codes as row names and years as column names. If NULL, standard errors are set to zero.}

\item{min_year}{An \code{integer}. First year of time-series.}

\item{max_year}{An \code{integer}. Last year of time-series.}

\item{species_name_ordre}{A \verb{character vector}. Species code in the same order as in \code{data_ts}.}

\item{species_sub}{A \code{data.frame} of species names. It should be provided with species in row, the first column for complete species names and the second column for species names' codes.}

\item{nfac}{An \code{integer}. Number of trends for the DFA. Default is 0 to test several values (between \code{mintrend} and \code{maxtrend}).}

\item{mintrend}{An \code{integer}. Minimum number of trends to test. Default is 1.}

\item{maxtrend}{An \code{integer}. Minimum number of trends to test. Default is 5.}

\item{AIC}{A \code{logical} value. \code{TRUE} computes and displays the AIC, \code{FALSE} does not. Default is \code{TRUE}.}

\item{center_option}{An \code{integer}. Option to handle time-series centered according to the first year (\code{center_option} = 0) or mean centred (default, \code{center_option} = 1).}

\item{silent}{A \code{logical} value. \code{TRUE} silences \code{MakeADFun()}, \code{FALSE} does not. Default is \code{TRUE}.}

\item{control}{A \code{list}. Control option for \code{MakeADFun()}. Default is list().}
}
\value{
A \code{fit_dfa} object composed of a list of 21 objects: \code{tmbObj} the output of \code{MakeADFun()}, \code{tmbOpt} the optimisation from \code{tmbObj}, \code{data_ts} the processed \code{matrix} of species time series, \code{data_ts_se} the processed \code{matrix} of time series uncertainty, \code{data_ts_save} a \code{data.frame} of input species time series, \code{data_ts_save_long} a \code{data.frame} of input species time series in long format, \code{data_ts_se_save} a \code{data.frame} of input species uncertainty time series, \code{nfac} the number of latent trends, \code{ny} the number of time series, \code{nT} the number of time steps, \code{aic} the value of the AIC, \code{conv} the result of the convergence check, \code{sdRep_test} the summary of the TMB optimisation output, \code{sdRep_test_all} the complete TMB optimisation output, \code{x_hat} estimated latent trends, \code{x_hat_se} standard error of estimated latent trends, \code{Z_hat} estimated factor loadings, \code{Z_hat_orig} initial estimated factor loadings, \code{Z_hat_se} standard error of estimated factor loadings, \code{cov_mat_Z} covariance matrix of factor loadings, \code{data_loadings} \code{data.frame} of factor loadings.
}
\description{
Run DFA
}
\examples{
\dontrun{
data(species_ts_mat)
data(species_uncert_ts_mat)

data_ready_dfa <- prepare_data(data_ts = species_ts_mat,data_ts_se = species_uncert_ts_mat,
se_log = TRUE, perc_replace = 0.01)

dfa_result <- fit_dfa(data_ts = data_ready_dfa$data_ts,data_ts_se = data_ready_dfa$data_ts_se,
min_year = data_ready_dfa$min_year, max_year = data_ready_dfa$max_year, species_name_ordre = data_ready_dfa$species_name_ordre,
species_sub = species_name, nfac = 0, mintrend = 1, maxtrend = 5, AIC = TRUE,
center_option = 1, silent = TRUE, control = list())
}
}
