% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_make_dfa.R
\name{fun_make_dfa}
\alias{fun_make_dfa}
\title{Run DFA}
\usage{
fun_make_dfa(
  data_ts,
  data_ts_se,
  min_year,
  max_year,
  species_name_ordre,
  species_sub,
  nfac = 0,
  mintrend = 1,
  maxtrend = 5,
  AIC = TRUE,
  center_option = 1,
  silent = TRUE,
  control = list()
)
}
\arguments{
\item{data_ts}{A \code{matrix} of species time series. It should be provided as a \code{matrix} with species time-series in row, species' codes as row names and years as column names.}

\item{data_ts_se}{A \code{matrix} of species of uncertainty (e.g. standard error) of species time series. It should be provided as a \code{matrix} with log species uncertainty time-series in row, species' codes as row names and years as column names.}

\item{min_year}{An \code{integer}. First year of time-series.}

\item{max_year}{An \code{integer}. Last year of time-series.}

\item{species_name_ordre}{A \verb{character vector}. Species code following the order of \code{data_ts}.}

\item{species_sub}{A \code{data.frame} of species names. It should be provided with species in row, the first column for complete species names and the second column for species names' codes.}

\item{nfac}{An \code{integer}. Number of trends for the DFA. Default is 0 to test several values (between \code{mintrend} and \code{maxtrend}).}

\item{mintrend}{An \code{integer}. Minimum number of trends to test. Default is 1.}

\item{maxtrend}{An \code{integer}. Minimum number of trends to test. Default is 5.}

\item{AIC}{A \code{logical} value. \code{TRUE} computes and displays the AIC, \code{FALSE} does not. Default is \code{TRUE}.}

\item{center_option}{An \code{integer}. Option to handle time-series centered according to the first year (\code{center_option} = 0) or mean centred (default, \code{center_option} = 1).}

\item{silent}{A \code{logical} value. \code{TRUE} silences \code{MakeADFun()}, \code{FALSE} does not. Default is \code{TRUE}.}

\item{control}{A \code{list}. Control option for \code{MakeADFun()}. Default is list().}
}
\value{
A list of 17 objects: \code{data_to_plot_sp} data on species time-series and fit, \code{data_to_plot_tr} data on latent trends, \code{data_loadings} data on factor loadings, \code{exp_var_lt} data on \% of variance of species ts explained by latent trends, \code{plot_sp} plot of species time-series and fit, \code{plot_tr} plot of latent trends, \code{plot_ld} plot of factor loadings, \code{plot_perc_var} plot of \% of variance of species ts explained by latent trends, \code{plot_sp_group} plot clusters in factorial plan, \code{plot_group_ts} plot clustertime-series, \code{plot_group_ts2} plot clustertime-series from sdRep, \code{aic} best AIC, \code{sdRep} optimisation output, \code{group_dfa} cluster results, \code{trend_group} cluster centre times-series, \code{trend_group2} cluster centre times-series from sdRep, \code{data_msi} data for multi-species index
}
\description{
Run DFA
}
\examples{
data(species_ts_mat)
data(species_uncert_ts_mat)
species_ts_mat[species_ts_mat==0] <- NA

data_ready_dfa <- data_check_prepare(data_ts = species_ts_mat,data_ts_se = species_uncert_ts_mat,
se_log = FALSE,is_mean_centred = FALSE, min_year_sc=2000)

dfa_result <- fun_make_dfa(data_ts = data_ready_dfa$data_ts,data_ts_se = data_ready_dfa$data_ts_se,
min_year = data_ready_dfa$min_year, max_year = data_ready_dfa$max_year, species_name_ordre = data_ready_dfa$species_name_ordre,
species_sub = species_name, nfac = 0, mintrend = 1, maxtrend = 5, AIC = TRUE,
center_option = 1, silent = TRUE, control = list())
}
