% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_from_dfa_boot1.R
\name{group_from_dfa_boot1}
\alias{group_from_dfa_boot1}
\title{Get cluster for DFA results}
\usage{
group_from_dfa_boot1(
  data_loadings,
  cov_mat_Z,
  species_sub,
  nboot = 100,
  ny,
  nfac
)
}
\arguments{
\item{data_loadings}{A \code{data.frame} of species initial factor loadings.}

\item{cov_mat_Z}{A \code{matrix} corresponding to the covariance matrix of species factor loadings.}

\item{species_sub}{A \code{data.frame} of species names. It should be provided with species in row, the first column for complete species names and the second column for species names' codes.}

\item{nboot}{An \code{integer}. Number of bootstrap iteration. Default is 100.}

\item{ny}{An \code{integer}. Number of time series.}

\item{nfac}{An \code{integer}. Number of latent trends.}
}
\value{
A \code{list} of five objects: \code{kmeans_res} a data.frame containing the results of clustering, \code{centroids} a data.frame with the coordinates of cluster centres, \code{stability_cluster_final} a vector of stability for each cluster, \code{mean_dist_clust} a vector of the mean distance between species and centre for each cluster, \code{pca_centre_list} a list containing coordinates to plot trends of PCA axes.
}
\description{
Get cluster for DFA results
}
\examples{
\dontrun{
group_from_dfa_boot1(data_loadings, cov_mat_Z, species_sub, nboot=nboot, ny, nfac)
}
}
