% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dfa.R
\name{make_dfa}
\alias{make_dfa}
\title{Combine all functions}
\usage{
make_dfa(
  data_ts,
  data_ts_se,
  nfac = 0,
  mintrend = 1,
  maxtrend = 5,
  AIC = TRUE,
  species_sub,
  nboot = 500,
  silent = TRUE,
  control = list(),
  se_log = TRUE,
  is_mean_centred = TRUE,
  min_year_sc = NULL
)
}
\arguments{
\item{data_ts}{A \code{data.table}. Dataset of species time series. It should be provided as a data.table with species time-series in row, the first column for species names' codes and years as column names.}

\item{data_ts_se}{A \code{data.table}. Dataset of uncertainty (e.g. standard error) of species time series. It should be provided as a data.table with log species uncertainty time-series in row, the first column for species names' codes and years as column names.}

\item{nfac}{A \code{integer}. Number of trends for the DFA. Default is 0 to test several values (between \code{mintrend} and \code{maxtrend}).}

\item{mintrend}{An \code{integer}. Minimum number of trends to test. Default is 1.}

\item{maxtrend}{An \code{integer}. Minimum number of trends to test. Default is 5.}

\item{AIC}{A \code{logical} value. \code{TRUE} computes and displays the AIC, \code{FALSE} does not. Default is \code{TRUE}.}

\item{species_sub}{A \code{data.frame} of species names. It should be provided with species in row, the first column for complete species names and the second column for species names' codes.}

\item{nboot}{An \code{integer}. Number of bootstrap iteration. Default is 500.}

\item{silent}{A \code{logical} value. \code{TRUE} silences \code{MakeADFun()}, \code{FALSE} does not. Default is \code{TRUE}.}

\item{control}{A \code{list}. Control option for \code{MakeADFun()}. Default is list().}

\item{se_log}{A \code{logical} value. \code{TRUE} if \code{data_ts_se} is provided with log values. \code{FALSE} if it is provided with not log values. Default is \code{TRUE}.}

\item{is_mean_centred}{A \code{logical} value. \code{TRUE} if \code{data_ts} is provided with mean-centred values. \code{FALSE} if it is provided with non mean-centred values. Default is \code{TRUE}.}

\item{min_year_sc}{An \code{integer}. First year of the period to select for rescaling if \code{is_mean_centred} = \code{FALSE}.}
}
\value{
A list of 17 objects: \code{data_to_plot_sp} data on species time-series and fit, \code{data_to_plot_tr} data on latent trends, \code{data_loadings} data on factor loadings, \code{exp_var_lt} data on \% of variance of species ts explained by latent trends, \code{plot_sp} plot of species time-series and fit, \code{plot_tr} plot of latent trends, \code{plot_ld} plot of factor loadings, \code{plot_perc_var} plot of \% of variance of species ts explained by latent trends, \code{plot_sp_group} plot clusters in factorial plan, \code{plot_group_ts} plot clustertime-series, \code{plot_group_ts2} plot clustertime-series from sdRep, \code{aic} best AIC, \code{sdRep} optimisation output, \code{group_dfa} cluster results, \code{trend_group} cluster centre times-series, \code{trend_group2} cluster centre times-series from sdRep, \code{data_msi} data for multi-species index
}
\description{
Combine all functions
}
\examples{
\dontrun{
data(species_ts)
data(species_uncert_ts)
data(species_name)
species_ts[species_ts==0] <- NA
dfa_aus_farm <- make_dfa(data_ts = species_ts,data_ts_se = species_uncert_ts,
species_sub = species_name,nfac = 0,
mintrend = 1,maxtrend = 5,AIC = TRUE,
nboot = 10,silent = TRUE,control = list(),
se_log = FALSE,is_mean_centred = FALSE, min_year_sc=2000)
}
}
