% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NbClustb.R
\name{NbClustb}
\alias{NbClustb}
\title{Clustering function removing the buggy Beale index from NbClust}
\usage{
NbClustb(
  data = NULL,
  diss = NULL,
  distance = "euclidean",
  min.nc = 2,
  max.nc = 15,
  method = NULL,
  index = "all",
  alphaBeale = 0.1
)
}
\arguments{
\item{data}{matrix or dataset.}

\item{diss}{dissimilarity matrix to be used. By default, diss=NULL, but if it is replaced by a dissimilarity matrix, distance should be "NULL".}

\item{distance}{the distance measure to be used to compute the dissimilarity matrix. This must be one of: "euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski" or "NULL". By default, distance="euclidean". If the distance is "NULL", the dissimilarity matrix (diss) should be given by the user. If distance is not "NULL", the dissimilarity matrix should be "NULL".}

\item{min.nc}{minimal number of clusters, between 1 and (number of objects - 1)}

\item{max.nc}{maximal number of clusters, between 2 and (number of objects - 1), greater or equal to min.nc. By default, max.nc=15.}

\item{method}{the cluster analysis method to be used. This should be one of: "ward.D", "ward.D2", "single", "complete", "average", "mcquitty", "median", "centroid", "kmeans".}

\item{index}{the index to be calculated. This should be one of : "kl", "ch", "hartigan", "ccc", "scott", "marriot", "trcovw", "tracew", "friedman", "rubin", "cindex", "db", "silhouette", "duda", "pseudot2", "ratkowsky", "ball", "ptbiserial", "gap", "frey", "mcclain", "gamma", "gplus", "tau", "dunn", "hubert", "sdindex", "dindex", "sdbw", "all" (all indices except GAP, Gamma, Gplus and Tau), "alllong" (all indices with Gap, Gamma, Gplus and Tau included).}

\item{alphaBeale}{significance value for Beale's index.}
}
\value{
All.index Values of indices for each partition of the dataset obtained with a number of clusters between min.nc and max.nc.
All.CriticalValues	Critical values of some indices for each partition obtained with a number of clusters between min.nc and max.nc.
Best.nc Best number of clusters proposed by each index and the corresponding index value.
Best.partition	Partition that corresponds to the best number of clusters
}
\description{
Clustering function removing the buggy Beale index from NbClust
}
