% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dfa_result.R
\name{plot_dfa_result}
\alias{plot_dfa_result}
\title{Plot DFAclust results}
\usage{
plot_dfa_result(
  data_dfa,
  sdRep,
  species_sub,
  group_dfa,
  min_year,
  species_name_ordre
)
}
\arguments{
\item{data_dfa}{A \code{fit_dfa} object. Results of \code{fit_dfa()}.}

\item{sdRep}{A \code{TMB} object. Summary of the TMB optimisation output.}

\item{species_sub}{A \code{data.frame} of species names. It should be provided with species in row, the first column for complete species names and the second column for species names' codes.}

\item{group_dfa}{A \code{list}. Clustering output.}

\item{min_year}{An \code{integer}. First year of time-series.}

\item{species_name_ordre}{A \verb{character vector}. Species code following the order of \code{data_ts}.}
}
\value{
A \code{list} of 13 objects: \code{data_to_plot_sp} data on species time-series and fit, \code{data_to_plot_tr} Data on latent trends, \code{data_loadings} Data on factor loadings, \code{exp_var_lt} Data on \% of variance of species ts explained by latent trends, \code{plot_sp} Plot of species time-series and fit, \code{plot_tr} Plot of latent trends, \code{plot_ld} Plot of factor loadings, \code{plot_perc_var} Plot of \% of variance of species ts explained by latent trends, \code{plot_sp_group} Plot clusters in factorial plan, \code{plot_group_ts} Plot clustertime-series, \code{plot_group_ts2} Plot clustertime-series from sdRep, \code{trend_group} Cluster barycentre times-series, \code{trend_group2} Cluster barycentre times-series from sdRep.
}
\description{
Plot DFAclust results
}
\examples{
data(species_ts_mat)
data(species_uncert_ts_mat)

data_ready_dfa <- prepare_data(data_ts = species_ts_mat,data_ts_se = species_uncert_ts_mat,
se_log = TRUE, perc_replace = 0.01)

dfa_result <- fit_dfa(data_ts = data_ready_dfa$data_ts,data_ts_se = data_ready_dfa$data_ts_se,
min_year = data_ready_dfa$min_year, max_year = data_ready_dfa$max_year, species_name_ordre = data_ready_dfa$species_name_ordre,
species_sub = species_name, nfac = 0, mintrend = 1, maxtrend = 5, AIC = TRUE,
center_option = 1, silent = TRUE, control = list())

data(species_name)

cluster_result <- cluster_dfa(data_dfa = dfa_result, species_sub = species_name, nboot = 500)

dfa_result_plot <- plot_dfa_result(data_dfa = dfa_result, sdRep = cluster_result$sdRep,
species_sub = species_name, group_dfa = cluster_result$group_dfa,
min_year = data_ready_dfa$min_year, species_name_ordre = data_ready_dfa$species_name_ordre)
}
