% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Function to check data before DFA}
\usage{
prepare_data(data_ts, data_ts_se, se_log = TRUE, perc_replace = 0.01)
}
\arguments{
\item{data_ts}{A \code{matrix}. Dataset of species time series. It should be provided as a matrix with species time-series in row, species' codes as row names and years as column names.}

\item{data_ts_se}{A \code{matrix}. Dataset of uncertainty (e.g. standard error) of species time series. It should be provided as a matrix with log uncertainty of species time-series in row, species' codes as row names and years as column names. If NULL, standard errors are set to zero.}

\item{se_log}{A \code{logical} value. \code{TRUE} if \code{data_ts_se} is provided with log values. \code{FALSE} if it is provided with not log values. Default is \code{TRUE}.}

\item{perc_replace}{A \code{numeric} value. Proportion of the average index value used to replace zeros. Default is \code{0.01}.}
}
\value{
A \code{list} of five objects: \code{min_year} an \code{integer} corresponding to the first year of time-series, \code{max_year} an \code{integer} corresponding to the last year of time-series, \code{species_name_ordre} a \verb{character vector} of species code following the ordre of \code{data_ts}, \code{data_ts} a \code{matrix} with ready-to-use time-series for DFA (\code{fun_make_dfa()}), \code{data_ts_se} a \code{matrix} with ready-to-use uncertainty for DFA (\code{fun_make_dfa()}).
}
\description{
Function to check data before DFA
}
\examples{
data(species_ts_mat)
data(species_uncert_ts_mat)

data_ready_dfa <- prepare_data(data_ts = species_ts_mat,data_ts_se = species_uncert_ts_mat,
se_log = TRUE, perc_replace = 0.01)
}
