% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_index.R
\name{rescale_index}
\alias{rescale_index}
\title{Log mean-centred species indices and uncertainty}
\usage{
rescale_index(index, se, ref, denom = 1.01)
}
\arguments{
\item{index}{A numeric vector containing indices of a species across years.}

\item{se}{A numeric vector containing uncertainty that comes with indices of a species across years.}

\item{ref}{A logical vector containing the years to be used at reference for the scaling (TRUE) and the years not used in the reference (FALSE).}

\item{denom}{A numeric value inferior to 1 and superior to 0 to calculate the variance of the first year.}
}
\value{
A matrix/array with log mean-centred indices and uncertainty
}
\description{
Log mean-centred species indices and uncertainty
}
\examples{
set.seed(123)
index <- c(100,round(runif(19, min = 50, max = 150)))
se <- c(0,round(runif(19, min = 5, max = 15)))
ref <- c(rep(FALSE,5),rep(TRUE,15))
rescale_index(index, se, ref, denom = 1.01)
}
