% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data3.R
\docType{data}
\name{species_name}
\alias{species_name}
\title{Species Names and Codes}
\format{
\subsection{\code{species_name}}{

A data.table with 21 rows and 2 columns:
\describe{
\item{name_long}{Latin name}
\item{code_sp}{Species code}
}
}
}
\source{
\url{https://zenodo.org/record/4590199/files/national_indices2017.csv?download=1}
}
\usage{
species_name
}
\description{
Link species names and codes.
}
\keyword{datasets}
