% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1.R
\docType{data}
\name{species_ts_mat}
\alias{species_ts_mat}
\title{Species Indices of Farmland Bird in Austria}
\format{
\subsection{\code{species_ts_mat}}{

A matrix with 21 rows and 20 columns:
\describe{
\item{row.names}{Species code}
\item{1998 to 2017}{Years}
}
}
}
\source{
\url{https://zenodo.org/record/4590199/files/national_indices2017.csv?download=1}
}
\usage{
species_ts_mat
}
\description{
A subset of data from the Pan European Common Monitoring Bird Scheme
}
\keyword{datasets}
