% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{species_uncert_ts}
\alias{species_uncert_ts}
\title{Uncertainty of Species Indices of Farmland Bird in Austria}
\format{
\subsection{\code{species_uncert_ts}}{

A data.table with 21 rows and 20 columns:
\describe{
\item{code_sp}{Species code}
\item{1999 to 2017}{Years}
}
}
}
\source{
\url{https://zenodo.org/record/4590199/files/national_indices2017.csv?download=1}
}
\usage{
species_uncert_ts
}
\description{
A subset of data from the Pan European Common Monitoring Bird Scheme
}
\keyword{datasets}
