% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_dfa.R
\name{update_dfa}
\alias{update_dfa}
\title{Updating DFA parameters after clustering}
\usage{
update_dfa(tmbObj, Z_pred_from_kmeans, W_from_kmeans)
}
\arguments{
\item{tmbObj}{A \code{MakeADFun} object. Output of \code{MakeADFun()}.}

\item{Z_pred_from_kmeans}{A \code{matrix}. Factor loadings of cluster centres.}

\item{W_from_kmeans}{A \code{matrix}.  Weight of species time series.}
}
\value{
A \code{list} of two objects: \code{tmbObj} updated output of \code{MakeADFun()}, \code{sdRep} updated output of TMB optimisation
}
\description{
Updating DFA parameters after clustering
}
\examples{
data(species_ts_mat)
data(species_uncert_ts_mat)
species_ts_mat[species_ts_mat==0] <- NA

data_ready_dfa <- data_check_prepare(data_ts = species_ts_mat,data_ts_se = species_uncert_ts_mat,
se_log = FALSE,is_mean_centred = FALSE, min_year_sc=2000)

dfa_result <- fun_make_dfa(data_ts = data_ready_dfa$data_ts,data_ts_se = data_ready_dfa$data_ts_se,
min_year = data_ready_dfa$min_year, max_year = data_ready_dfa$max_year, species_name_ordre = data_ready_dfa$species_name_ordre,
species_sub = species_name, nfac = 0, mintrend = 1, maxtrend = 5, AIC = TRUE,
center_option = 1, silent = TRUE, control = list())

data(species_name)

cluster_result <- cluster_dfa(data_loadings = dfa_result$data_loadings,
cov_mat_Z = dfa_result$cov_mat_Z, species_sub = species_name,
nboot = 500, ny = dfa_result$ny, nfac = dfa_result$nfac,
data_ts = dfa_result$data_ts)

dfa_result_update <- update_dfa(tmbObj = dfa_result$tmbObj,
Z_pred_from_kmeans = cluster_result$Z_pred_from_kmeans,
W_from_kmeans = cluster_result$W_from_kmeans)
}
