% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_dfa.R
\name{cluster_dfa}
\alias{cluster_dfa}
\title{Clustering function to process DFA results.}
\usage{
cluster_dfa(
  data_loadings,
  cov_mat_Z,
  species_sub,
  nboot = 500,
  ny,
  nfac,
  data_ts
)
}
\arguments{
\item{data_loadings}{A \code{data.frame}. Factor loadings and standard error.}

\item{cov_mat_Z}{A \code{matrix}. Covariance matrix of factor loadings.}

\item{species_sub}{A \code{data.frame} of species names. It should be provided with species in row, the first column for complete species names and the second column for species names' codes.}

\item{nboot}{An \code{integer}. Number of bootdtrap iterations for clustering. Default is 500.}

\item{ny}{An \code{integer}. Number of time-series.}

\item{nfac}{An \code{integer}. Number of latent trends.}

\item{data_ts}{A \code{matrix} of species time series.}
}
\value{
A \code{list} of three objects: \code{group_dfa} clustering output, \code{Z_pred_from_kmeans} factor loadings of clusters, \code{W_from_kmeans} weight of species time series.
}
\description{
Clustering function to process DFA results.
}
\examples{
\dontrun{
data(species_ts_mat)
data(species_uncert_ts_mat)
species_ts_mat[species_ts_mat==0] <- NA

data_ready_dfa <- data_check_prepare(data_ts = species_ts_mat,data_ts_se = species_uncert_ts_mat,
se_log = FALSE,is_mean_centred = FALSE, min_year_sc=2000)

dfa_result <- fun_make_dfa(data_ts = data_ready_dfa$data_ts,data_ts_se = data_ready_dfa$data_ts_se,
min_year = data_ready_dfa$min_year, max_year = data_ready_dfa$max_year, species_name_ordre = data_ready_dfa$species_name_ordre,
species_sub = species_name, nfac = 0, mintrend = 1, maxtrend = 5, AIC = TRUE,
center_option = 1, silent = TRUE, control = list())

data(species_name)

cluster_result <- cluster_dfa(data_loadings = dfa_result$data_loadings,
cov_mat_Z = dfa_result$cov_mat_Z, species_sub = species_name, nboot = 500,
ny = dfa_result$ny, nfac = dfa_result$nfac, data_ts = dfa_result$data_ts)
}
}
