% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dfa_result.R
\name{plot_dfa_result}
\alias{plot_dfa_result}
\title{Plot DFAclust results}
\usage{
plot_dfa_result(
  data_ts_save,
  data_ts_se_save,
  data_ts,
  data_ts_se,
  sdRep,
  ny,
  species_sub,
  x_hat,
  x_hat_se,
  Z_hat,
  nfac,
  group_dfa,
  nT,
  min_year
)
}
\arguments{
\item{data_ts_save}{A \code{data.frame}. Input species time series.}

\item{data_ts_se_save}{A \code{data.frame}. Input uncertainty of species time series.}

\item{data_ts}{A \code{matrix}. The processed \code{matrix} of species time series.}

\item{data_ts_se}{A \code{matrix}. The processed \code{matrix} of uncertainty of species time series.}

\item{sdRep}{A \code{TMB} object. Summary of the TMB optimisation output.}

\item{ny}{An \code{integer}. Number of time series.}

\item{species_sub}{A \code{data.frame} of species names. It should be provided with species in row, the first column for complete species names and the second column for species names' codes.}

\item{x_hat}{A \code{matrix}. Estimated latent trends.}

\item{x_hat_se}{A \code{matrix}. Standard error of estimated latent trends.}

\item{Z_hat}{A \code{matrix}. Estimated factor loadings.}

\item{nfac}{An \code{integer}. Number of latent trends.}

\item{group_dfa}{A \code{list}. Clustering output.}

\item{nT}{An \code{integer}. Number of time steps.}

\item{min_year}{An \code{integer}. First year of time-series.}
}
\value{
A \code{list} of 13 objects: \code{data_to_plot_sp} data on species time-series and fit, \code{data_to_plot_tr} Data on latent trends, \code{data_loadings} Data on factor loadings, \code{exp_var_lt} Data on \% of variance of species ts explained by latent trends, \code{plot_sp} Plot of species time-series and fit, \code{plot_tr} Plot of latent trends, \code{plot_ld} Plot of factor loadings, \code{plot_perc_var} Plot of \% of variance of species ts explained by latent trends, \code{plot_sp_group} Plot clusters in factorial plan, \code{plot_group_ts} Plot clustertime-series, \code{plot_group_ts2} Plot clustertime-series from sdRep, \code{trend_group} Cluster barycentre times-series, \code{trend_group2} Cluster barycentre times-series from sdRep.
}
\description{
Plot DFAclust results
}
\examples{
\dontrun{
data(species_ts_mat)
data(species_uncert_ts_mat)
species_ts_mat[species_ts_mat==0] <- NA

data_ready_dfa <- data_check_prepare(data_ts = species_ts_mat,data_ts_se = species_uncert_ts_mat,
se_log = FALSE,is_mean_centred = FALSE, min_year_sc=2000)

dfa_result <- fun_make_dfa(data_ts = data_ready_dfa$data_ts,data_ts_se = data_ready_dfa$data_ts_se,
min_year = data_ready_dfa$min_year, max_year = data_ready_dfa$max_year, species_name_ordre = data_ready_dfa$species_name_ordre,
species_sub = species_name, nfac = 0, mintrend = 1, maxtrend = 5, AIC = TRUE,
center_option = 1, silent = TRUE, control = list())

data(species_name)

cluster_result <- cluster_dfa(data_loadings = dfa_result$data_loadings,
cov_mat_Z = dfa_result$cov_mat_Z, species_sub = species_name,
nboot = 500, ny = dfa_result$ny, nfac = dfa_result$nfac,
data_ts = dfa_result$data_ts)

dfa_result_update <- update_dfa(tmbObj = dfa_result$tmbObj,
Z_pred_from_kmeans = cluster_result$Z_pred_from_kmeans,
W_from_kmeans = cluster_result$W_from_kmeans)

dfa_result_plot = plot_dfa_result(data_ts_save = dfa_result$data_ts_save, data_ts_se_save = dfa_result$data_ts_se_save,
data_ts = dfa_result=data_ts, data_ts_se = dfa_result$data_ts_se, sdRep = dfa_result_update$sdRep,
ny = dfa_result$ny, species_sub = species_name, x_hat = dfa_result$x_hat, x_hat_se = dfa_result$x_hat_se,
Z_hat = dfa_result$Z_hat, nfac = dfa_result$nfac, group_dfa = cluster_result$group_dfa,
nT = dfa_result$nT, min_year = data_ready_dfa$min_year)
}
}
