% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_group_boot.R
\name{plot_group_boot}
\alias{plot_group_boot}
\title{Plot groups and clustering trends}
\usage{
plot_group_boot(
  nb_group,
  centroids,
  kmeans_res,
  sdrep,
  nT,
  min_year,
  stability_cluster_final,
  mean_dist_clust,
  pca_centre,
  Z_hat,
  x_hat,
  data_ts,
  data_ts_se,
  data_to_plot_sp
)
}
\arguments{
\item{nb_group}{Number of clusters}

\item{centroids}{Position of cluster barycentres}

\item{kmeans_res}{Results of clustering}

\item{sdrep}{Optimisation output from DFA}

\item{nT}{Number of time step}

\item{min_year}{Oldest year in time-series}

\item{stability_cluster_final}{Stability of clusters}

\item{mean_dist_clust}{Average distance between species and barycentre}

\item{pca_centre}{Coordinates to plot trends of PCA axes}

\item{Z_hat}{Factor loadings}

\item{x_hat}{Latent trends}

\item{data_ts}{A \code{data.table}. Dataset of species time series. It should be provided as a data.table with species time-series in row, the first column for species names' codes and years as column names.}

\item{data_ts_se}{A \code{data.table}. Dataset of uncertainty (e.g. standard error) of species time series. It should be provided as a data.table with species uncertainty time-series in row, the first column for species names' codes and years as column names.}

\item{data_to_plot_sp}{Data for species time-series plot}
}
\value{
A \code{list} of six objects: \code{final_plot_list} the plot of species clusters, \code{graph} the plot of time-series of cluster centres, \code{data_trend_group} the \code{data.frame} of time-series of cluster centres, \code{graph2} the plot of time-series of cluster centres from sdRep, \code{data_trend_group2} the \code{data.frame} of time-series of cluster centres from sdRep, \code{data_msi} the \code{data.frame} for multi-species index
}
\description{
Plot groups and clustering trends
}
\examples{
\dontrun{
plot_group_boot(nb_group = nrow(group_dfa[[1]][[2]]), centroids = group_dfa[[2]], kmeans_res = group_dfa[[1]], sdrep = sdRep, nT = nT, min_year = min_year, stability_cluster_final = group_dfa[[3]], mean_dist_clust = group_dfa[[4]], pca_centre = group_dfa[[5]],Z_hat = Z_hat,x_hat = x_hat,data_ts = data_ts,data_ts_se = data_ts_se,data_to_plot_sp = data_to_plot_sp)
}
}
